+++ src/pymetropolis/metro_calibration/road/penalties.py
@@ -4,18 +4,13 @@ from loguru import logger
 
 from pymetropolis.metro_common.errors import MetropyError
 from pymetropolis.metro_common.io import read_dataframe
+from pymetropolis.metro_network.road_network.common import default_edge_values_validator
+from pymetropolis.metro_network.road_network.files import RoadEdgesCleanFile, RoadEdgesUrbanFlagFile
 from pymetropolis.metro_pipeline import Step
 from pymetropolis.metro_pipeline.parameters import CustomParameter, PathParameter
 from pymetropolis.metro_pipeline.steps import InputFile
 
-from .common import default_edge_values_validator
-from .files import (
-    RoadEdgesCleanFile,
-    RoadEdgesFreeFlowTravelTimeFile,
-    RoadEdgesPenaltiesFile,
-    RoadEdgesUrbanFlagFile,
-    RoadEdgesVariablesFile,
-)
+from .files import RoadEdgesFreeFlowTravelTimeFile, RoadEdgesPenaltiesFile, RoadEdgesVariablesFile
 
 if TYPE_CHECKING:
     import geopandas as gpd